move functions to utils/mapUtils file
